pyproject.toml
[tool.pytest.ini_options]
addopts = "--timeout=900 --json-report-file=test_failure_report.json --json-report-indent=4 --json-report-omit=environment --json-report-omit=collectors --json-report --json-report-omit=keywords"

[tool.poetry]
name = "portability"
version = "0.1.0"
description = ""
authors = ["frasermince"]

[tool.poetry.group.m1]
optional = true

[tool.poetry.group.intel]
optional = true

[tool.poetry.group.vm]
optional = true

[tool.poetry.group.tpuvm]
optional = true

[tool.poetry.group.torch]
optional = true


[tool.poetry.group.m1.dependencies]
code-tokenize = "^0.2.0"
tensorflow-macos = "~2.11.0"

[tool.poetry.group.intel.dependencies]
code-tokenize = "^0.2.0"
tensorflow = "~2.11.0"

[tool.poetry.group.vm.dependencies]
tensorflow = "~2.11.0"

[tool.poetry.group.tpuvm.dependencies]
tensorflow = "~2.11.0"
cloud-tpu-client = "^0.10"

[tool.poetry.group.torch.dependencies]
torch = "~1.12.0"



[tool.poetry.dependencies]
python = ">=3.8,<3.11"
datasets = "^2.4.0"
tqdm = "^4.64.0"
beautifulsoup4 = "^4.11.1"
matplotlib = "^3.6.0"
numpy = "^1.23.3"
seaborn = "^0.12.0"
pandas = "^1.5.0"
pytest = "^7.1.3"
absl-py = "^1.3.0"
protobuf = "^3.19"
expecttest = "^0.1.4"
hypothesis = "^6.58.0"
cmake = "^3.13.3"
pytest-xdist = "^3.2.1"
pytest-forked = "^1.6.0"
portpicker = "^1.5.2"
scipy = "^1.10.1"
pytest-timeout = "^2.1.0"
pytest-json-report = "^1.5.0"
jax = "^0.4.7"
jaxlib = "^0.4.7"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
